Upgrade to Astro 5 + supporting deps
- Upgrade `astro` to 5.x
- Upgrade `astro-embed` to remove peer dep warning
- Drop `@astrojs/tailwind`
- Upgrade Tailwind to 4.x
- Upgrade `@astrojs/solid-js` to 5.x to support Astro 5
- Upgrade `@astrojs/mdx` to 4.x to support Astro 5
- Replace content collections with Content Layer API
