+++ package.json
@@ -13,17 +13,17 @@
     "start": "astro dev"
   },
   "dependencies": {
-    "@astrojs/check": "^0.7.0",
-    "@astrojs/mdx": "^3.1.1",
-    "@astrojs/solid-js": "^4.4.0",
-    "@astrojs/tailwind": "^5.1.0",
-    "@astrojs/ts-plugin": "^1.8.0",
+    "@astrojs/check": "0.9.6",
+    "@astrojs/mdx": "^4.3.13",
+    "@astrojs/solid-js": "^5.1.3",
+    "@astrojs/ts-plugin": "1.10.6",
     "@fontsource-variable/figtree": "^5.0.21",
     "@fontsource/ibm-plex-mono": "^5.0.13",
     "@iconify-json/heroicons": "^1.1.20",
     "@shikijs/transformers": "^1.3.0",
-    "astro": "^4.11.0",
-    "astro-embed": "^0.7.1",
+    "@tailwindcss/vite": "^4.1.18",
+    "astro": "^5.16.6",
+    "astro-embed": "^0.9.2",
     "astro-icon": "^1.1.0",
     "classnames": "^2.5.1",
     "hast-util-from-html": "^2.0.1",
@@ -37,7 +37,8 @@
     "remark-directive": "^3.0.0",
     "remark-math": "^6.0.0",
     "solid-js": "^1.8.17",
-    "tailwindcss": "^3.4.3",
+    "tailwind-merge": "^3.4.0",
+    "tailwindcss": "^4.1.18",
     "throttleit": "^2.1.0",
     "unist-util-visit": "^5.0.0"
   },
@@ -54,7 +55,8 @@
     "prettier-plugin-astro": "^0.13.0",
     "prettier-plugin-packagejson": "^2.5.0",
     "typescript": "^5.4.5",
-    "vercel": "^34.1.8"
+    "vercel": "^34.1.8",
+    "vite": "^6.4.1"
   },
   "packageManager": "yarn@4.1.0",
   "lavamoat": {
