+++ src/components/GalleryComponent.tsx
