+++ src/components/LinkOrNoLink.astro
@@ -1,6 +1,4 @@
 ---
-import Image from "./Image.astro";
-
 type Props = {
   linkToRawVersion: boolean;
   href: string;
