--- src/env.d.ts
@@ -1,2 +1,0 @@
-/// <reference path="../.astro/types.d.ts" />
-/// <reference types="astro/client" />
