+++ src/hooks/useScreenWidth.ts
@@ -1,20 +1,7 @@
 import throttle from "throttleit";
-import resolveConfig from "tailwindcss/resolveConfig";
-import tailwindConfig from "../../tailwind.config.mjs"; // Fix the path
 import { createSignal, onCleanup, onMount } from "solid-js";
 
-const fullConfig = resolveConfig(tailwindConfig);
-const screens = fullConfig.theme.screens;
-type Screens = typeof screens;
-type Breakpoints = { [K in keyof Screens]: number };
-const breakpoints = Object.entries(screens).reduce<Breakpoints>(
-  (obj, [name, value]) => {
-    return { ...obj, [name]: value.replace(/px$/, "") };
-  },
-  {} as Breakpoints,
-);
-
-export function useTailwindBreakpoints() {
+export function useScreenWidth() {
   const [screenWidth, setScreenWidth] = createSignal<number>(0);
 
   const onResize = throttle(() => {
@@ -34,5 +21,5 @@ export function useTailwindBreakpoints() {
     }
   });
 
-  return { breakpoints, screenWidth };
+  return { screenWidth };
 }
