+++ tsconfig.json
@@ -12,6 +12,8 @@
     {
       "name": "@astrojs/ts-plugin"
     }
-  ]
+  ],
+  "include": [".astro/types.d.ts", "**/*"],
+  "exclude": ["dist"]
 }
